+++ apis/python/src/tiledbsoma/_dataframe.py
@@ -409,7 +409,7 @@ class DataFrame(SOMAArray, somacore.DataFrame):
     @property
     def tiledbsoma_has_upgraded_domain(self) -> bool:
         """Returns true if the array has the upgraded resizeable domain feature
-        from TileDB-SOMA 1.14: the array was created with this support, or it has
+        from TileDB-SOMA 1.15: the array was created with this support, or it has
         had ``.tiledbsoma_upgrade_domain`` applied to it.
 
         Lifecycle: