+++ apis/python/src/tiledbsoma/_tdb_handles.py
@@ -485,7 +485,7 @@ class DataFrameWrapper(SOMAArrayWrapper[clib.SOMADataFrame]):
     @property
     def tiledbsoma_has_upgraded_domain(self) -> bool:
         """Returns true if the array has the upgraded resizeable domain feature
-        from TileDB-SOMA 1.14: the array was created with this support, or it has
+        from TileDB-SOMA 1.15: the array was created with this support, or it has
         had ``.tiledbsoma_upgrade_domain`` applied to it.
 
         Lifecycle:
@@ -502,7 +502,7 @@ class DenseNDArrayWrapper(SOMAArrayWrapper[clib.SOMADenseNDArray]):
     @property
     def tiledbsoma_has_upgraded_shape(self) -> bool:
         """Returns true if the array has the upgraded resizeable shape feature
-        from TileDB-SOMA 1.14: the array was created with this support, or it has
+        from TileDB-SOMA 1.15: the array was created with this support, or it has
         had ``.tiledbsoma_upgrade_shape`` applied to it.
 
         Lifecycle:
@@ -531,7 +531,7 @@ class SparseNDArrayWrapper(SOMAArrayWrapper[clib.SOMASparseNDArray]):
     @property
     def tiledbsoma_has_upgraded_shape(self) -> bool:
         """Returns true if the array has the upgraded resizeable shape feature
-        from TileDB-SOMA 1.14: the array was created with this support, or it has
+        from TileDB-SOMA 1.15: the array was created with this support, or it has
         had ``.tiledbsoma_upgrade_shape`` applied to it.
 
         Lifecycle: