+++ apis/python/src/tiledbsoma/options/_tiledb_create_write_options.py
@@ -157,9 +157,6 @@ class TileDBCreateOptions:
     attrs: Mapping[str, _ColumnConfig] = attrs_.field(
         factory=dict, converter=_normalize_columns
     )
-    consolidate_and_vacuum: bool = attrs_.field(
-        validator=vld.instance_of(bool), default=False
-    )
 
     @classmethod
     def from_platform_config(
@@ -199,7 +196,7 @@ class TileDBCreateOptions:
         """Constructs the real TileDB Filters to use for offsets."""
         assert_version_before(1, 14)
         warnings.warn(
-            "`offsets_filters_tiledb` is now deprecated for removal in 1.14 "
+            "`offsets_filters_tiledb` is now deprecated for removal in 1.15 "
             "as we no longer support returning tiledb.Filter. "
             "Use `offsets_filters` instead.",
             DeprecationWarning,
@@ -211,7 +208,7 @@ class TileDBCreateOptions:
         """Constructs the real TileDB Filters to use for the validity map."""
         assert_version_before(1, 14)
         warnings.warn(
-            "`validity_filters_tiledb` is now deprecated for removal in 1.14 "
+            "`validity_filters_tiledb` is now deprecated for removal in 1.15 "
             "as we no longer support returning tiledb.Filter. "
             "Use `validity_filters` instead.",
             DeprecationWarning,
@@ -226,7 +223,7 @@ class TileDBCreateOptions:
         """Constructs the real TileDB Filters to use for the named dimension."""
         assert_version_before(1, 14)
         warnings.warn(
-            "`dim_filters_tiledb` is now deprecated for removal in 1.14 "
+            "`dim_filters_tiledb` is now deprecated for removal in 1.15 "
             "as we no longer support returning tiledb.Filter. "
             "Use `dims` instead.",
             DeprecationWarning,
@@ -247,7 +244,7 @@ class TileDBCreateOptions:
         """Constructs the real TileDB Filters to use for the named attribute."""
         assert_version_before(1, 14)
         warnings.warn(
-            "`attr_filters_tiledb` is now deprecated for removal in 1.14 "
+            "`attr_filters_tiledb` is now deprecated for removal in 1.15 "
             "as we no longer support returning tiledb.Filter. "
             "Use `attrs` instead.",
             DeprecationWarning,