+++ apis/r/R/Init.R
@@ -25,10 +25,6 @@
     } else {
       .pkgenv[["use_current_domain_transitional_internal_only"]] <- FALSE
     }
-    if (rpkg_lib_version != soma_lib_version) {
-      msg <- sprintf("TileDB Core version %s used by TileDB-R package, but TileDB-SOMA uses %s [%s]",
-                     sQuote(rpkg_lib_version), sQuote(soma_lib_version), sQuote(cdmsg))
-    }
 }
 
 # This is temporary only. Please see:
@@ -62,6 +58,9 @@
 #' @param config A named character vector with \sQuote{key} and \sQuote{value} pairs defining the
 #' configuration setting
 #' @return An external pointer object containing a shared pointer instance of \code{SOMAContext}
+#' @examples
+#' cfgvec <- as.vector(tiledb::tiledb_config())   # TileDB Config in vector form
+#' sctx <- soma_context(cfgvec)
 #' @export
 soma_context <- function(config) {